+++ src/views/system/role/list.vue
@@ -1,16 +1,37 @@
 <template>
   <div class="app-container">
     <div class="filter-container">
-      <el-input v-model="roleName" :placeholder="'角色名'" style="width: 200px;" class="filter-item"
-        @keyup.enter.native="handleFilter" />
-      <el-button class="filter-item" type="primary" icon="el-icon-search" style="margin-left:10px;"
-        @click="handleFilter">
-        {{ $t('table.search') }}</el-button>
-      <el-button class="filter-item" style="margin-left: 10px;" type="primary" icon="el-icon-edit"
-        @click="handleCreate">{{ $t('table.add') }}</el-button>
+      <el-input
+        v-model="roleName"
+        :placeholder="'角色名'"
+        style="width: 200px;"
+        class="filter-item"
+        @keyup.enter.native="handleFilter"
+      />
+      <el-button
+        class="filter-item"
+        type="primary"
+        icon="el-icon-search"
+        style="margin-left:10px;"
+        @click="handleFilter"
+      >{{ $t('table.search') }}</el-button>
+      <el-button
+        class="filter-item"
+        style="margin-left: 10px;"
+        type="primary"
+        icon="el-icon-edit"
+        @click="handleCreate"
+      >{{ $t('table.add') }}</el-button>
     </div>
 
-    <el-table v-loading="listLoading" :data="list" border fit highlight-current-row style="width: 100%">
+    <el-table
+      v-loading="listLoading"
+      :data="list"
+      border
+      fit
+      highlight-current-row
+      style="width: 100%"
+    >
       <el-table-column align="center" label="ID" width="80">
         <template slot-scope="scope">
           <span>{{ scope.row.Id }}</span>
@@ -27,7 +48,9 @@
 
       <el-table-column class-name="status-col" label="Enabled" width="110">
         <template slot-scope="scope">
-          <el-tag :type="scope.row.Enabled ? 'success' : 'danger'">{{ scope.row.Enabled ? '正常' : '禁用' }}</el-tag>
+          <el-tag
+            :type="scope.row.Enabled ? 'success' : 'danger'"
+          >{{ scope.row.Enabled ? '正常' : '禁用' }}</el-tag>
         </template>
       </el-table-column>
 
@@ -43,19 +66,31 @@
             <el-button type="primary" size="small" icon="el-icon-edit">{{ $t('table.edit') }}</el-button>
           </router-link>
           <router-link :to="'/system/role/assign/'+scope.row.Id">
-            <el-button type="warning" size="small" icon="el-icon-share" style="margin-left: 5px;">
-              {{ $t('table.assign') }}</el-button>
+            <el-button
+              type="warning"
+              size="small"
+              icon="el-icon-share"
+              style="margin-left: 5px;"
+            >{{ $t('table.assign') }}</el-button>
           </router-link>
-          <el-button size="small" type="danger" icon="el-icon-delete" style="margin-left: 5px;"
-            @click="handleDelete(scope.row)">
-            {{ $t('table.delete') }}
-          </el-button>
+          <el-button
+            size="small"
+            type="danger"
+            icon="el-icon-delete"
+            style="margin-left: 5px;"
+            @click="handleDelete(scope.row)"
+          >{{ $t('table.delete') }}</el-button>
         </template>
       </el-table-column>
     </el-table>
 
-    <pagination v-show="total>0" :total="total" :page.sync="listQuery.page" :limit.sync="listQuery.limit"
-      @pagination="getList" />
+    <pagination
+      v-show="total>0"
+      :total="total"
+      :page.sync="listQuery.page"
+      :limit.sync="listQuery.limit"
+      @pagination="getList"
+    />
   </div>
 </template>
 
@@ -67,8 +102,7 @@ import Pagination from "@/components/Pagination"; // Secondary package based on 
 export default {
   name: "RoleList",
   components: { Pagination },
-  filters: {
-  },
+  filters: {},
   data() {
     return {
       list: null,
@@ -91,13 +125,15 @@ export default {
     };
   },
   watch: {
-    roleName: function (newVal, oldVal) {
-      let conditions = [{
-        Field: "IsDeleted",
-        DataType: util.query.dataType.boolean,
-        Option: util.query.opt.eq,
-        Value: false
-      }];
+    roleName: function(newVal, oldVal) {
+      let conditions = [
+        {
+          Field: "IsDeleted",
+          DataType: util.query.dataType.boolean,
+          Option: util.query.opt.eq,
+          Value: false
+        }
+      ];
 
       if (newVal != "") {
         conditions.push({
@@ -136,25 +172,31 @@ export default {
       this.$router.push({ path: "/system/role/create" });
     },
     handleDelete(row) {
-      deleteRole(row.Id).then(response => {
-        let { success, msg } = response.data;
-        if (success) {
-          this.$notify({
-            title: '提示',
-            message: '删除成功',
-            type: 'success',
-            duration: 2000
-          })
-          this.getList();
-        } else {
-          this.$notify({
-            title: '提示',
-            message: msg,
-            type: 'error',
-            duration: 2000
-          })
-        }
-      })
+      this.$confirm("您确定要执行该操作吗？", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      }).then(() => {
+        deleteRole(row.Id).then(response => {
+          let { success, msg } = response.data;
+          if (success) {
+            this.$notify({
+              title: "提示",
+              message: "删除成功",
+              type: "success",
+              duration: 2000
+            });
+            this.getList();
+          } else {
+            this.$notify({
+              title: "提示",
+              message: msg,
+              type: "error",
+              duration: 2000
+            });
+          }
+        });
+      });
     },
     handleFilter() {
       this.listQuery.page = 1;
